appointment transaction fix, pdf all data fetch

+++ lib/screens/lawyer_screen/screens/models/lawyer_appoint_trans_model.dart
@@ -55,7 +55,7 @@ class Data {
   String? paymentMethod;
   double? paymentAmount;
   String? paymentStatus;
-  int? likhitDeCommission;
+  double? likhitDeCommission;
   double? gatewayCommission;
   double? splitCharge;
   double? likhitDeNetAmt;
